+++ CHANGELOG.md
@@ -44,6 +44,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed bugs
 
+* `jj git init --colocate` now refuses to run inside a Git worktree, providing
+  a helpful error message with alternatives.
+  [#8052](https://github.com/jj-vcs/jj/issues/8052)
+
 * `jj git push` now ensures that tracked remote bookmarks are updated even if
   there are no mappings in the Git fetch refspecs.
   [#5115](https://github.com/jj-vcs/jj/issues/5115)